Fixed some scrapers and added missing function.json file for azure function to pickup. Improvement in base_scraper.
