+++ CanadaShelters/__init__.py
@@ -16,10 +16,6 @@ def main(mytimer: func.TimerRequest, context: func.Context) -> None:
     ).isoformat()
     conn_string = os.environ['MONGO_DB_CONNECTION_STRING']
     client = MongoClient(conn_string)['shelter']
-    if stored_update_date is not None:
-        if scraped_update_date < stored_update_date:
-            logging.info('No new data. Goodbye...')
-            sys.exit()
     CSS.main_scraper(client)
     if mytimer.past_due:
         logging.info('The timer is past due!')
