--- HUD_scrapper/readme.md
@@ -1,1 +1,0 @@
-
