+++ LittleHelpBookData/function.json
@@ -1,0 +1,11 @@
+{
+    "scriptFile": "__init__.py",
+    "bindings": [
+      {
+        "name": "mytimer",
+        "type": "timerTrigger",
+        "direction": "in",
+        "schedule": "0 0 15 1-7 * Fri"
+      }
+    ]
+}
