+++ LittleHelpBookData/little_help_book.py
@@ -16,7 +16,7 @@ if _i not in sys.path:
 del _i  # clean up global name space
 from shared_code.utils import (
     check_similarity, locate_potential_duplicate,
-    insert_services, client
+    insert_services, get_mongo_client
 )
 from shared_code.base_scraper import BaseScraper
 
@@ -96,4 +96,5 @@ lhb_scraper = LHBScraper(
 )
 
 if __name__ == '__main__':
+    client = get_mongo_client()
     lhb_scraper.main_scraper(client)
