+++ NWHospitality/__init__.py
@@ -16,14 +16,7 @@ def main(mytimer: func.TimerRequest, context: func.Context) -> None:
     ).isoformat()
     conn_string = os.environ['MONGO_DB_CONNECTION_STRING']
     client = MongoClient(conn_string)['shelter']
-    scraped_update_date = nwh_scraper.scrape_updated_date()
-    stored_update_date = nwh_scraper.retrieve_last_scraped_date(client)
-
-    if stored_update_date is not None:
-        if scraped_update_date < stored_update_date:
-            logging.info('No new NWH data. Goodbye...')
-            sys.exit()
     nwh_scraper.main_scraper(client)
     if mytimer.past_due:
         logging.info('The timer is past due!')
-    logging.info(f'Python timer trigger function for NWH Scraping ran at utc: {utc_timestamp}')
+    logging.info(f'Python timer trigger function for North West Hospitality Scraping ran at utc: {utc_timestamp}')
