+++ NYCRunawayAndHomelessYouth/nyc_runaway_and_homeless_youth.py
@@ -16,7 +16,7 @@ if _i not in sys.path:
 del _i  # clean up global name space
 from shared_code.utils import (
     check_similarity, locate_potential_duplicate,
-    insert_services, client
+    insert_services, get_mongo_client
 )
 from shared_code.base_scraper import BaseScraper
 
@@ -69,6 +69,7 @@ nyc_youth_scraper = NYCYouthScraper(
 
 
 if __name__ == '__main__':
+    client = get_mongo_client()
     nyc_youth_scraper.main_scraper(client)
