--- VeteranCenters/__init__.py
@@ -1,21 +1,0 @@
-import logging
-import json
-import os
-import sys
-
-import datetime
-from pymongo import MongoClient, TEXT
-import azure.functions as func
-
-from .Veteran_Centers import veteran_shelters_scraper
-
-def main(mytimer: func.TimerRequest, context: func.Context) -> None:
-    utc_timestamp = datetime.datetime.utcnow().replace(
-        tzinfo=datetime.timezone.utc
-    ).isoformat()
-    conn_string = os.environ['MONGO_DB_CONNECTION_STRING']
-    client = MongoClient(conn_string)['shelter']
-    washington_dc_shelters_scraper.main_scraper(client)
-    if mytimer.past_due:
-        logging.info('The timer is past due!')
-    logging.info(f'Python timer trigger function for Veteran Centers scraping ran at utc: {utc_timestamp}')
